Using a cast to a float instead of using the f notation.

+++ src/chrisperkins/training/Main.java
@@ -15,7 +15,7 @@ public class Main {
         System.out.println("Double maximum value: " + myMaxDoubleValue);
 
         int myIntValue = 5;
-        float myFloatValue = 5.25f;
+        float myFloatValue = (float) 5.25;
         double myDoubleValue = 5d;
         
     }
